+++ src/main/java/net/dracoblue/spring/web/mvc/method/annotation/HttpResponseHeader.java
@@ -17,4 +17,14 @@ public @interface HttpResponseHeader {
 	 * 
 	 */
 	String value() default "";
+	
+	/**
+	 * 
+	 */
+	boolean valueExpression() default false;
+
+	/**
+	 * 
+	 */
+	boolean nameExpression() default false;
 }