+++ src/test/java/net/dracoblue/spring/web/mvc/method/annotation/NotAnnotatedTestController.java
@@ -29,8 +29,7 @@ public class NotAnnotatedTestController
 			}
 		)
 	public String handleWithEmptyMultipleExtraAnnotatedMethod() {
-		return null;
-		
+		return null;	
 	}
 	
 }